+++ src/components/functional/toasts-list/components/transaction-toast/components/transaction-toast-content/transaction-toast-content.tsx
@@ -8,6 +8,8 @@ import type { ITransactionListItem } from 'components/visual/transaction-list-it
 import { DataTestIdsEnum } from 'constants/dataTestIds.enum';
 
 import type { IToastDataState } from '../../transaction-toast.type';
+import { TransactionStatusEnum } from 'constants/transactionStatus.enum';
+import { getIsTransactionFailed } from 'utils/getTransactionStatus';
 
 // prettier-ignore
 const styles = {
@@ -35,9 +37,6 @@ export class TransactionToastContent {
     const showAmount = this.transactions.length === 1 && transaction?.amount;
     const showExplorerLinkButton = transaction?.link && this.transactions.length === 1;
     const amount = transaction && getAmountParts(transaction.amount);
-    const isBatch = this.transactions.length > 1;
-    const showPrimaryIcon = !isBatch && (
-      transaction.asset == null || transaction.asset.imageUrl || transaction.asset.icon || transaction.asset.text);
     const showTooltip = showAmount && amount.label.length > 10;
 
     return (
@@ -49,7 +48,7 @@ export class TransactionToastContent {
         data-testid={DataTestIdsEnum.transactionToastContent}
       >
         <div class="transaction-toast-content">
-          {!showPrimaryIcon && this.toastDataState.icon ? (
+          {this.toastDataState.icon ? (
             <Icon
               name={this.toastDataState.icon}
               class={classNames('transaction-toast-icon',
@@ -63,9 +62,9 @@ export class TransactionToastContent {
           ) : (
             <div
               class={classNames('transaction-toast-icon', {
-                'transaction-toast-icon-failed': transaction.status === 'fail' || transaction.status === 'invalid',
-                'transaction-toast-icon-pending': transaction.status === 'pending',
-                'transaction-toast-icon-success': transaction.status === 'success',
+                'transaction-toast-icon-failed': getIsTransactionFailed(transaction.status),
+                'transaction-toast-icon-pending': transaction.status === TransactionStatusEnum.pending,
+                'transaction-toast-icon-success': transaction.status === TransactionStatusEnum.success,
               })}
             >
               <TransactionAssetIcon transaction={transaction} iconSize={IconSizeEnumType.small} />
@@ -95,7 +94,7 @@ export class TransactionToastContent {
                   class={classNames('transaction-toast-amount', {
                     'amount-negative': transaction.amount.startsWith('-'),
                     'amount-positive': !transaction.amount.startsWith('-'),
-                    'transaction-toast-failed': transaction.status === 'fail' || transaction.status === 'invalid',
+                    'transaction-toast-failed': getIsTransactionFailed(transaction.status)
                   })}
                   isValid
                   label={amount.label}