+++ src/components/visual/transaction-list-item/transaction-list-item.tsx
@@ -4,6 +4,7 @@ import { IconSizeEnumType, TransactionAssetIcon } from 'common/TransactionAssetI
 import { getAmountParts } from 'components/functional/toasts-list/helpers';
 
 import type { ITransactionListItem } from './transaction-list-item.types';
+import { getIsTransactionFailed } from 'utils/getTransactionStatus';
 
 @Component({
   tag: 'mvx-transaction-list-item',
@@ -24,7 +25,7 @@ export class TransactionListItem {
         <div class="transaction-item">
           <div
             class={classNames('transaction-icon', {
-              'transaction-icon-failed': ['fail', 'invalid'].includes(this.transaction.status),
+              'transaction-icon-failed': getIsTransactionFailed(this.transaction.status),
             })}
           >
             <TransactionAssetIcon transaction={this.transaction} iconSize={IconSizeEnumType.large} />
@@ -49,7 +50,7 @@ export class TransactionListItem {
                   class={classNames('transaction-amount', {
                     'amount-negative': this.transaction.amount.startsWith('-'),
                     'amount-positive': !this.transaction.amount.startsWith('-'),
-                    'transaction-failed': this.transaction.status === 'fail' || this.transaction.status === 'invalid',
+                    'transaction-failed': getIsTransactionFailed(this.transaction.status),
                   })}
                   isValid
                   label={amount.label}